Ensure we read streams fully
Read method makes an assumption that the stream will return the
amount of data specified by the length parameter which isn't the case for
streams.

The length parameter indicates how much it may read up to. A stream will
block until it has read at least one byte, after that it may return at
any point indicating how much it actually read.

Also fixing a bug in the EntryInputStream which made the same assumption
that the number of bytes read will always be the same as the length
requested.